requirements.txt
torch>=2.2
torchvision>=0.17
torchaudio>=2.2
tqdm>=4.66
numpy>=1.26
scikit-learn>=1.4
opencv-python>=4.10
matplotlib>=3.9
pyyaml>=6.0
albumentations>=1.4
einops>=0.8
